requirements.txt
attrs==21.4.0
Authlib==0.15.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
requests==2.27.1
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
